size_t for read_field

+++ benchmarks/benchmark-charpoly.C
@@ -39,7 +39,7 @@ int main(int argc, char** argv) {
   
 	size_t iter = 1;
 	int    q    = 131071;
-	int    n    = 2000;
+	size_t    n    = 2000;
 	std::string file = "";
   	static int variant =0;
